refactor(spike): drop the ModalDefault shim, repoint helpers to the controller
interceptClose/setInterceptClose/setModalTitle callers now import from the
controller directly; the ModalDefault module and its obsolete Storybook story
are removed. No reactstrap remains in the modal-portal path. (Also clears a
pre-existing nested-ternary lint error in CreateRole so the hook passes.)
