+++ frontend/web/components/modals/CreateProject.tsx
@@ -13,7 +13,7 @@ import { RouteComponentProps } from 'react-router-dom'
 import ErrorMessage from 'components/ErrorMessage'
 import Button from 'components/base/forms/Button'
 import InputGroup from 'components/base/forms/InputGroup'
-import { setInterceptClose } from './base/ModalDefault'
+import { setInterceptClose } from './base/modalController'
 import PlanBasedAccess from 'components/PlanBasedAccess'
 import UserSelect from 'components/UserSelect'
 import MyRoleSelect from 'components/MyRoleSelect'
