+++ frontend/web/components/modals/CreateSegment.tsx
@@ -40,7 +40,7 @@ import { cloneDeep } from 'lodash'
 import ProjectStore from 'common/stores/project-store'
 import AddMetadataToEntity from 'components/metadata/AddMetadataToEntity'
 import { useGetSupportedContentTypeQuery } from 'common/services/useSupportedContentType'
-import { setInterceptClose } from './base/ModalDefault'
+import { setInterceptClose } from './base/modalController'
 import SegmentRuleDivider from 'components/SegmentRuleDivider'
 import { useGetProjectQuery } from 'common/services/useProject'
 import { useCreateProjectChangeRequestMutation } from 'common/services/useProjectChangeRequest'
