+++ frontend/web/components/modals/create-feature/index.tsx
@@ -20,7 +20,7 @@ import TabItem from 'components/navigation/TabMenu/TabItem'
 import ChangeRequestModal from 'components/modals/ChangeRequestModal'
 import classNames from 'classnames'
 import { useHasPermission } from 'common/providers/Permission'
-import { setInterceptClose } from 'components/modals/base/ModalDefault'
+import { setInterceptClose } from 'components/modals/base/modalController'
 import { getStore } from 'common/store'
 import ExternalResourcesTable from 'components/ExternalResourcesTable'
 import GitHubLinkSection from 'components/GitHubLinkSection'
